merge a bit from an iOS list
https://github.com/BlackJack8/iOSAdblockList

I removed many of the DMCA takedown domains, dead hosts, some of the web click-type domains trying to get it a little more narrowed down to mobile ad networks.  We had some overlap of several hundred mobile-specific domains so it seemed like a good fit.  Also, many mobile ad companies target both platforms so I wanted to catch that.  Additionally, this user is in another part of the world so the types of ad networks he may encounter might be different than the ones I do.