scrub of much of Disconnect.me Malvertising
most of this stuff is browser-based and most other browser ad blockers catch them already.  I kept some that were either familiar as mobile ad library or that felt like they might be.
For some reason I'd also blocked icann.com which is silly.  I must've imported it in with one of the bulk lists because that's awful.