big cleanup to bring it more in line with mobile
Over the course of importing a few other lists and additions by myself it's got a bit out of focus.  This is meant to be a mobile in-app ad/tracker blocklist for those 3rd parties that offer monetization via sdk code embedded in apps.
It still isn't perfect but this looks better.  I took the well-known EasyList and pulled out all domains that were in the format ||example.com^$third-party.  As EasyList represents ads collected by browsing the net, this takes much of that stuff out.

So, if you want browser blocklist, use browser adblockers with the regular blocklists out there.  I also want this to focus on what other lists do not have so it is adds a particular value and might be more appealing to merge than simply another list for browser-based traffic.